+++ core/src/main/java/io/questdb/griffin/engine/groupby/TimestampSampler.java
@@ -6,7 +6,7 @@
  *    \__\_\\__,_|\___||___/\__|____/|____/
  *
  *  Copyright (c) 2014-2019 Appsicle
- *  Copyright (c) 2019-2024 QuestDB
+ *  Copyright (c) 2019-2026 QuestDB
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -36,7 +36,7 @@ public interface TimestampSampler extends Sinkable {
 
     int getTimestampType();
 
-    long nextTimestamp(long timestamp, int numSteps);
+    long nextTimestamp(long timestamp, long numSteps);
 
     default long nextTimestamp(long timestamp) {
         return nextTimestamp(timestamp, 1);
